+++ frontend/src/components/Buttons/Button.tsx
@@ -1,5 +1,5 @@
 import  { MouseEvent, ElementType } from 'react';
-import { defaultButtonStyle, errorButtonStyle } from '../styles/ButtonStyles.tsx'
+import { defaultButtonStyle, errorButtonStyle } from './ButtonStyles.tsx'
 
 type ButtonProps = {
     text?: string;
