+++ frontend/src/pages/SignInPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import Button from "../components/Button";
+import Button from "../components/Buttons/Button";
 import TextInput from "../components/TextInput";
 import Form from "../components/Form";
 import Header from "../components/Header";
